Turn random blocks red when they're "scanned"
The way the block mesh is mapped to the "bad" block mesh is kind of a
hack as the GridMap doesn't provide a convenient way to do that, but it
works.

Had to raise the viewer and change the way the scanner looks to make
this look right. Also made the grid narrower to make sure plenty of red
blocks appear inside our FOV.